% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toPhyloseq.R
\name{addSampleData}
\alias{addSampleData}
\title{addSampleData}
\usage{
addSampleData(MA, sampleData = NULL)
}
\arguments{
\item{MA}{A \code{\link{MultiAmplicon}} object}

\item{sampleData}{A data frame of providing data for samples in
the \code{\link{MultiAmplicon}} object. If set to \code{NULL}
(default) sampleData will be added based on names of the
\code{link{PairedReadFileSet}} and the resulting sampleData
will (only) give names of forward and reverse file names for
each sample.}
}
\value{
A \code{\link{MultiAmplicon}} object with the sampleData
    slot filled.
}
\description{
Add sample data to a MultiAmplicon object
}
\details{
The sampleData slot is filled with a
\code{\link[phyloseq]{sample_data}} object from phyoseq created
merging sample names (colnames) of the MultiAmplicon object with a
data frame of sample data. The rownames of the that sampleData
data frame must correspond to colnames of the MultiAmplcion
object.
}
\author{
Emanuel Heitlinger
}
