% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carnivoreSeqRuns}
\alias{carnivoreSeqRuns}
\title{Sequencing run files for the carnivore biome dataset}
\format{An object of class \code{data.frame} with 192 rows and 5 columns.}
\source{
\url{https://www.ncbi.nlm.nih.gov/bioproject/PRJNA386767}
}
\usage{
carnivoreSeqRuns
}
\description{
A dataset containing SRA submission, details of the SRA dataset for
BioProject PRJNA386767
"Intestinal biome sequencing of carnivores". This data was
generated in R using the following code:
}
\examples{
## showing how this was generated.
\dontrun{sqlfile <- "SRAmetadb.sqlite"
if(!file.exists(sqlfile)){
    sqlfile <- getSRAdbFile()
}
sra_con <- dbConnect(SQLite(),sqlfile)
carnivoreSeqRuns <- getSRA(search_terms = '"Intestinal biome sequencing of carnivores"',
                          sra_con=sra_con, acc_only=TRUE)
}

}
\keyword{datasets}
