% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate-methods.R
\name{concatenateDadaMulti}
\alias{concatenateDadaMulti}
\title{concatenateDadaMulti}
\usage{
concatenateDadaMulti(MA1, MA2, what = "samples")
}
\arguments{
\item{MA1}{MultiAmplicon object that should be concatenated.}

\item{MA2}{Second MultiAmplicon object to be concatenated with the
first.}

\item{what}{Should either "samples" or "amplicons" be
concatenated? Currently only "samples" are implemented.}
}
\value{
A concatenated MultiAmplicon object
}
\description{
Concatenate two MultiAmplicon objects
}
\details{
Two MultiAmplicon objects are concatenated. Currently only
concatenation of different samples with the same amplicons is
implemented.
}
\author{
Emanuel Heitlinger
}
