% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{dadaMulti}
\alias{dadaMulti}
\title{dadaMulti}
\usage{
dadaMulti(MA, Ferr = NULL, Rerr = NULL, ...)
}
\arguments{
\item{MA}{MultiAmplicon-class object.}

\item{Ferr}{As in the "err" parameter of dada: the matrix of
estimated rates for each possible nucleotide change. In this
case for forward reads.}

\item{Rerr}{the same for the reverse reads.}

\item{...}{additional parameters to be passed to the
\code{\link[dada2]{dada}} function of \code{dada2}. All
arguments to dadaMulti can be given as a vector of the same
length as the number of primer pairs in the MultiAmplicon
object. This allows to specify different parameters for each
amplicon. If a shorter vector is given it will be recycled to
match the number of amplicons.}
}
\value{
MultiAmplicon object with dadaF and dadaR slots filled.
}
\description{
A wrapper around \code{\link[dada2]{dada}} of the \code{dada2}
packager for multiple amplicons.
}
\details{
The function runs \code{\link[dada2]{dada}} from the package
 \href{https://benjjneb.github.io/dada2/}{\code{dada2}} to perform
 'High resolution sample inference from amplicon data' on multiple
 amplicons stored as dereplicated sequences in a
 MultiAmplicon-class object.
}
\author{
Emanuel Heitlinger
}
