% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derepMulti.R
\name{derepMulti}
\alias{derepMulti}
\title{derepMulti}
\usage{
derepMulti(MA, mc.cores = getOption("mc.cores", 2L),
  keep.single.singlets = FALSE, ...)
}
\arguments{
\item{MA}{MultiAmplicon object to be dereplicated}

\item{mc.cores}{number or compute cores for parallel processing}

\item{keep.single.singlets}{logical argument indicationg whether a
derep slot should be filled with a single sequence recovered
in only one sample. Defaults to FALSE meaning that such empty
derep objects are created in such a case. Keeping such derep
objects (setting this to TRUE) might result in downstream
problems (Errors) in dada inference.}

\item{...}{arguments to be passed to \code{\link{derepFastq}}}
}
\value{
MultiAmplicon object with derep slots (forward derepF and
    reverse derepR) filled
}
\description{
Dereplicate sequences in fastq files
}
\details{
An interface to \code{\link[dada2]{derepFastq}} which itself uses
\code{\link[ShortRead]{FastqStreamer}} for dereplicating amplicon
sequences from fastq or compressed fastq files.
}
\author{
Emanuel Heitlinger
}
