% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toPhyloseq.R
\name{.fillSampleTables}
\alias{.fillSampleTables}
\title{.fillSampleTables}
\usage{
.fillSampleTables(ST, samples)
}
\arguments{
\item{ST}{a sequence tabel from a MultiAmplicon object}

\item{samples}{a character vector giving names of samples to
retain.}
}
\value{
a sequence table for use in a MultiAmplicon object
}
\description{
Fill sequence tables in a MultiAmplicon object to include selected
samples for all amplicons.
}
\details{
In a MultiAmplicon object for some primer pairs some samples might
have no amplified sequence variants at all. This function creates 
matrices including consistent samples for all the \code{sequenceTableNoChime}
slots of a MultiAmplicon object and returnes filled tables for all requested
samplese. Cells are filled with 0 (zeros) for samples originally
not recoverd in an amplicon.
}
\author{
Emanuel Heitlinger
}
