% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_track.R
\name{.sumThingsOrNull}
\alias{.sumThingsOrNull}
\title{.sumThingsOrNull}
\usage{
.sumThingsOrNull(y, what)
}
\arguments{
\item{y}{a vector of dada, derep, merger objects from dada2}
}
\value{
Numeric vector for the count of entries (uniques or
    sequences, ASVs)
}
\description{
Get a sum of the uniques vector or matrix or 0 if empty
}
\details{
This function uses getUniques from dada2 to extract (potentially)
unique entries (ASVs, sequences or merged sequences) from a vector
of suitable objects.
}
\author{
Emanuel Heitlinger
}
