% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{fillSampleTables}
\alias{fillSampleTables}
\title{fillSampleTables}
\usage{
fillSampleTables(MA, samples = "union")
}
\arguments{
\item{MA}{MultiAmplicon object}

\item{samples}{either a character vector of length one giving
"union" to use all samples appearing for any of the amplicons
or a longer character vector giving names of samples to
retain}
}
\value{
MultiAmplicon object with the \code{sequenceTableFilled}
    slot filled
}
\description{
Fill multiple sequence Tables in a MultiAmplicon object for all or
selected samples.
}
\details{
In a MultiAmplicon object for some primer pairs some samples might
have no amplified sequence variants at all. This function adds a
\code{sequenceTableFilled} slot to a MultiAmplicon object. It uses
the \code{sequenceTableNoChime} slot and fills tables for all
samples appearing in any table or for requested samples.
}
\author{
Emanuel Heitlinger
}
