% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_track.R
\name{getCounts}
\alias{getCounts}
\title{getCounts}
\usage{
getCounts(MaMat, what = "input")
}
\arguments{
\item{MaMat}{A matrix or vector of derep, dada or merger objects}
}
\value{
a matrix, list or vector of counts, depending on the class
    of the input
}
\description{
Get a sum of uniques from an input object.
}
\details{
This function calculates the number of unique sequences, ASVs or
merged sequences from the relevant slot in a MultiAmplicon object
}
\author{
Emanuel Heitlinger
}
