% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUnmatched.R
\name{getUnmatched}
\alias{getUnmatched}
\title{getUnmatched}
\usage{
getUnmatched(MA, tabulate.only = TRUE, outdir = tempdir())
}
\arguments{
\item{MA}{A code{\link{MultiAmplicon}} object on which
\code{link{sortAmplicons}} has been called previously.}

\item{tabulate.only}{Binary TRUE or FALSE indcating whether only
the number of matching reads should be tabulated if set to
false non-matched sequencing reads are written into outdir.}

\item{outdir}{output directory into which files for unmatched
reads are written. Files are unmatched_F.fastq and
unmatched_R.fastq for forward and reverse reads, respectively.}
}
\value{
A table indicting the number of read pairs matched (TRUE)
    and unmatched (FALSE) by primer pairs.
}
\description{
get the unmatched sequences for a MultiAmplicon project
}
\details{
This function computes the number of sequencing read pairs not
matched by primer pairs after running \code{\link{sortAmplicons}}
}
\author{
Emanuel Heitlinger
}
