% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{makeSequenceTableMulti}
\alias{makeSequenceTableMulti}
\title{sequenceTableMulti}
\usage{
makeSequenceTableMulti(MA, ...)
}
\arguments{
\item{MA}{\code{\link{MultiAmplicon-class}} object with mergers
slot filled}

\item{...}{additional parameters to be passed to the
\code{\link[dada2]{makeSequenceTable}} function of
\code{dada2}. All arguments to the function can be given as a
vector of the same length as the number of primer pairs in the
MultiAmplicon object, allowing to specify different parameters
for each amplicon. If a shorter vector is given it will be
recycled to match the number of amplicons.}
}
\value{
A MultiAmplicon-class object with the sequenceTable slot
    filled
}
\description{
Create a sequence table inside a MultiAmplicon object.
}
\details{
This is a wrapper for the \code{\link[dada2]{makeSequenceTable}}
function of \code{dada2}. It works on an
\code{\link{MultiAmplicon-class}} object with the mergers slot
filled. Use \code{\link{dadaMulti}} and \code{\link{derepMulti}}
on an amplicon sorted (see \code{\link{sortAmplicons}})
\code{\link{MultiAmplicon-class}} object to preprocess your
multi-marker data to this point.
}
\author{
Emanuel Heitlinger
}
