% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{MultiAmplicon-class}
\alias{MultiAmplicon-class}
\title{MultiAmplicon-class}
\usage{
## Constructors:

MultiAmplicon(PrimerPairsSet, PairedReadFileSet)
}
\arguments{
\item{PrimerPairsSet}{a set of primer pairs specifiying your
amplicons see \code{\link{PrimerPairsSet-class}}}

\item{PairedReadFileSet}{a set of paired end sequencing data fiels
\code{\link{PairedReadFileSet-class}}}
}
\value{
MultiAmplicon-class
}
\description{
The PrimerPairsSet class is a container for storing
    primer pairs.
}
\details{
A class combining sequences of forward and reverse primers (in a
\code{\link{PrimerPairsSet-class}}) plus file names of paired end
sequencing files (in a \code{\link{PairedReadFileSet-class}}).
}
\author{
Emanuel Heitlinger
}
\seealso{
\code{\link{DNAStringSet}}
}

