% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{noChimeMulti}
\alias{noChimeMulti}
\title{noChimeMulti}
\usage{
noChimeMulti(MA, mc.cores = getOption("mc.cores", 2L), ...)
}
\arguments{
\item{MA}{A \code{\link{MultiAmplicon-class}} object pre-processed
to have a sequenceTableMulti populated}

\item{mc.cores}{integer number of cores to use for parallelization}

\item{...}{paramter passed through to
\code{\link[dada2]{removeBimeraDenovo}}. All arguments to the
function can be given as a vector of the same length as the
number of primer pairs in the MultiAmplicon object, allowing
to specify different parameters for each amplicon. If a
shorter vector is given it will be recycled to match the
number of ampicons.}
}
\value{
a \code{\link{MultiAmplicon-class}} object with the
    sequenceTableNoChime filled
}
\description{
Remove chimeric sequencing read pairs from a MultiAmplicon
object.
}
\details{
This is a wrapper for the \code{\link[dada2]{removeBimeraDenovo}}
function of \code{dada2}. It works on an
\code{\link{MultiAmplicon-class}} object with the sequenceTable
slot filled. Use \code{\link{dadaMulti}},
\code{\link{derepMulti}}, \code{\link{mergeMulti}} and
\code{\link{sequenceTableMulti}} on a amplicon sorted (see
\code{\link{sortAmplicons}}) \code{\link{MultiAmplicon-class}}
object to preprocess your multi-marker data to this point.
}
\author{
Emanuel Heitlinger
}
