% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R, R/subset-methods.R
\docType{class}
\name{PairedDada-class}
\alias{PairedDada}
\alias{PairedDada-class}
\alias{[,PairedDada,ANY,ANY,ANY-method}
\alias{length,PairedDada-method}
\title{PairedDada-class}
\usage{
PairedDada(dadaF = list(), dadaR = list())

\S4method{length}{PairedDada}(x)

\S4method{[}{PairedDada,ANY,ANY,ANY}(x, i)
}
\description{
A pair of two dada objects
}
\details{
dada-class objects as defined by the package \code{dada2}
(function \code{\link[dada2]{dada}}) are bundeled as forward and
reverse read pairs in this object
}
\author{
Emanuel Heitlinger
}

