% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{rawCounts}
\alias{rawCounts}
\title{rawCounts}
\usage{
rawCounts(x)
}
\arguments{
\item{x}{A MultiAmplicon-class object}
}
\value{
A numerical matrix of counts for each amplicon
    (primer-pair) matched in each sample (paired read file)
}
\description{
Access rawCounts of a MultiAmplicon object
}
\details{
An accessor for the 'rawCounts' slot of a MultiAmplicon-class object
}
\author{
Emanuel Heitlinger
}

