% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{sequenceTableMulti}
\alias{sequenceTableMulti}
\title{sequenceTableMulti}
\usage{
sequenceTableMulti(MA, ...)
}
\arguments{
\item{MA}{\code{\link{MultiAmplicon-class}} object with mergers
slot filled}

\item{...}{additional parameters to be passed to the
\code{\link[dada2]{makeSequenceTable}} function of
\code{dada2}}
}
\value{
A MultiAmplicon-class object with the sequenceTable slot
    filled
}
\description{
Create a sequence table inside a MultiAmplcion object.
}
\details{
This is a wrapper for the \code{\link[dada2]{makeSequenceTable}}
function of \code{dada2}. It works on an
\code{\link{MultiAmplicon-class}} object with the mergers slot
filled. Use \code{\link{dadaMulti}} and \code{\link{derepMulti}}
on an amplicon sorted (see \code{\link{sortAmplicons}})
\code{\link{MultiAmplicon-class}} object to preprocess your
multi-marker data to this point.
}
\author{
Emanuel Heitlinger
}
