% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\docType{methods}
\name{show-methods}
\alias{show,MultiAmplicon-method}
\alias{show,PrimerPairsSet-method}
\alias{show-methods}
\title{show-methods}
\usage{
\S4method{show}{PairedReadFileSet}(object)

\S4method{show}{PrimerPairsSet}(object)

\S4method{show}{MultiAmplicon}(object)
}
\arguments{
\item{object}{A PairedReadFileSet-class object to be shown}

\item{object}{A PrimerPairsSet-class object to be shown}

\item{object}{A MultiAmplicon-class object to be shown}
}

