% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAmplicon-Methods.R
\docType{methods}
\name{sortAmplicons,MultiAmplicon-method}
\alias{sortAmplicons,MultiAmplicon-method}
\title{sortAmplicons}
\usage{
\S4method{sortAmplicons}{MultiAmplicon}(MA, n = 1e+06, ...)
}
\arguments{
\item{MA}{}

\item{n}{}

\item{...}{}
}
\value{
MultiAmplicon
}
\description{
.. content for description (no empty lines) ..
}
\details{
.. content for details
}
\author{
Emanuel Heitlinger
}

