% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-methods.R
\docType{methods}
\name{[,MultiAmplicon,ANY,ANY,logical-method}
\alias{[,MultiAmplicon,ANY,ANY,logical-method}
\title{subset MultiAmplicon}
\usage{
\S4method{[}{MultiAmplicon,ANY,ANY,logical}(x, i = TRUE, j = TRUE,
  drop = FALSE)
}
\arguments{
\item{i}{numeric, logical or names vector for subsetting rows (==
amplicons)}

\item{j}{numeric, logical or names vector for subsetting columns
(== read files, corresponding usually to samples)}

\item{drop}{should}

\item{MultiAmplicon-class}{object}
}
\value{
a subset of the original MultiAmplicon object
}
\description{
Convenient subsetting for MultiAmplicon objects
}
\details{
Subset a MultiAmplicon object including all potentially filled
slots
}
\author{
Emanuel Heitlinger
}

