% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toPhyloseq.R
\name{toPhyloseq}
\alias{toPhyloseq}
\title{toPhyloseq}
\usage{
toPhyloseq(MA, ...)
}
\arguments{
\item{MA}{MultiAmplicon object with the \code{taxonTable} and
\code{sequenceTableNoChime} slots filled.}
}
\value{
a \code{phyloseq} object or a list of such objects
}
\description{
Populate a phyloseq object for
}
\author{
Emanuel Heitlinger
}
