% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\docType{methods}
\name{calcPropMerged}
\alias{calcPropMerged}
\alias{calcPropMerged,MultiAmplicon-method}
\title{calcPropMerged}
\usage{
calcPropMerged(MA)

\S4method{calcPropMerged}{MultiAmplicon}(MA)
}
\arguments{
\item{MA}{MultiAmplicon object with the \code{mergers} slot
filled.}
}
\value{
a vector of proportions of merged reads for each amplicon
    (potentially named).
}
\description{
Calculate the proportion of merged sequences for a MultiAmplicon
object.
}
\details{
Merging using \code{\link{dadaMulti}} can result in loss of
sequences with too little overlap. This function determines what
proporton of sequences were merged using \code{\link{mergeMulti}}
for each amplicon in a MultiAmplicon object.
}
\author{
Emanuel Heitlinger
}
