% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dadaMulti.R
\name{dadaMulti}
\alias{dadaMulti}
\title{dadaMulti}
\usage{
dadaMulti(MA, ...)
}
\arguments{
\item{MA}{MultiAmplicon-class object}

\item{...}{additional parameters to be passed to
\code{\link{dada}} from \code{\link{dada}}}
}
\value{
MultiAmplicon object with dadaF and dadaR slots filled
}
\description{
A wrapper around \link{\code{dada}} from \link{\code{dada2}}
}
\details{
The function runs \link{\code{dada}} from \link{\code{dada2}} to
performe 'High resolution sample inference from amplicon data' on
multiple amplicons stored as dereplicated sequences in a
MultiAmplicon-class object
}
\author{
Emanuel Heitlinger
}

