% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dadaMulti.R
\name{dadaMulti}
\alias{dadaMulti}
\title{dadaMulti}
\usage{
dadaMulti(MA, ...)
}
\arguments{
\item{MA}{MultiAmplicon-class object}

\item{...}{additional parameters to be passed to
\code{\link[dada2]{dada}} from \code{\link[dada2]{dada}}}
}
\value{
MultiAmplicon object with dadaF and dadaR slots filled
}
\description{
A wrapper around \code{\link[dada2]{dada}} from \code{dada2}
}
\details{
The function runs \code{\link[dada2]{dada}} from the package
\code{dada2} \url{https://benjjneb.github.io/dada2/} to performe
'High resolution sample inference from amplicon data' on multiple
amplicons stored as dereplicated sequences in a
MultiAmplicon-class object
}
\author{
Emanuel Heitlinger
}
