% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dadaMulti.R
\name{dadaMulti}
\alias{dadaMulti}
\title{dadaMulti}
\usage{
dadaMulti(MA, ...)
}
\arguments{
\item{MA}{MultiAmplicon object}

\item{...}{Parameters to be passed to dada2::dada}
}
\value{
MultiAmplicon object with dadaF and dadaR slots filled
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Emanuel Heitlinger
}

