% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derepMulti.R
\name{derepMulti}
\alias{derepMulti}
\title{derepMulti}
\usage{
derepMulti(MA, ...)
}
\arguments{
\item{MA}{MultiAmplicon object to be dereplicated}

\item{...}{arguments to be passed to dada2::derepFastq}
}
\value{
MultiAmplicon object with derep slots filled
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Emanuel Heitlinger
}

