% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeMulti.R
\name{mergeMulti}
\alias{mergeMulti}
\title{mergeMulti}
\usage{
mergeMulti(MA, ...)
}
\arguments{
\item{MA}{\code{\link{MultiAmplicon-class}} object with derep and
dada slots filled.}

\item{...}{additional parameters to be passed to the
\code{\link[dada2]{mergePairs}} function of \code{dada2}}
}
\value{
A MultiAmplicon-class object with the mergers slot filled
}
\description{
merge denoised pairs of forward and reverse reads inside an
MultiAmplcion object.
}
\details{
This is a wrapper for \code{\link[dada2]{mergePairs}} from
\code{dada2}. It works on an \code{\link{MultiAmplicon-class}}
object with derep and dada slots filled. Use
\code{\link{dadaMulti}} and \code{\link{derepMulti}} on a amplicon
sorted (see \code{\link{sortAmplicons}})
\code{\link{MultiAmplicon-class}} object to preprocess your
multi-marker data to this point.
}
\author{
Emanuel Heitlinger
}
