% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{MultiAmplicon-class}
\alias{MultiAmplicon-class}
\title{MultiAmplicon-class}
\usage{
## Constructors:

MultiAmplicon(PrimerPairsSet, PairedReadFileSet)
}
\arguments{
\item{PrimerPairsSet}{a set of primer pairs specifiying your
amplicons see \code{\link{PrimerPairsSet-class}}}

\item{PairedReadFileSet}{a set of paired end sequencing data fiels
    \code{\link{PairedReadFileSet-class}}

Accessor-like methods:

In the code snippets below, 'x' is an MultiAmplicon-class object.

'nrow(x)' An integer giving the number of primer pairs in the
object

'ncol(x)' An integer giving the number of paired read files
(usually samples and their replicates) in the object

'rownames(x)' A character vector giving the primer-pair names see
\code{\link{PrimerPairsSet}}

'colnames(x)' A character vector giving the names of the paired
read filnames (e.g. used to store sample names) see also
\code{\link{PairedReadFileSet}}

'rawCounts(x)' A matrix of raw sequencing read counts assigned to
a particular sample (input file pair) and amplicon (input primer
pair)}
}
\value{
MultiAmplicon-class
}
\description{
A class combining sequences of forward and reverse primers (in a
\code{\link{PrimerPairsSet-class}}) plus file names of paired end
sequencing files (in a \code{\link{PairedReadFileSet-class}}).
}
\details{
The MultiAmplicon class is a container for storing primer pairs,
read files and processed data in an 'amplicon x samples'
format. Slots
}
\section{Slots}{

\describe{
\item{\code{PrimerPairsSet}}{The primer pairs used in your experiment to
specify amplicons stored in a PrimerPairsSet-class object.}

\item{\code{PairedReadFileSet}}{The (quality filtered) fastq files (one
file pair for each sample) that store your sequencing data.}

\item{\code{rawCounts}}{}

\item{\code{stratifiedFiles}}{temporary files as a result of stratifying
into amplicons and samples using the function
\code{\link{sortAmplicons}}. Forward (sometimes called R1) and
reverse (sometimes called R2) files are stored as a (amplicons
x samples) matrix of \code{\link{PairedReadFileSet-class}}
objects.}

\item{\code{derep}}{}

\item{\code{dada}}{}

\item{\code{mergers}}{}

\item{\code{sequenceTable}}{}

\item{\code{sequenceTableNoChime}}{}
}}
\author{
Emanuel Heitlinger
}
\seealso{
\code{\link{derep}},\code{\link{dada}}
}

