% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noChimeMulti.R
\name{noChimeMulti}
\alias{noChimeMulti}
\title{noChimeMulti}
\usage{
noChimeMulti(MA, mc.cores, ...)
}
\arguments{
\item{MA}{A \code{\link{MultiAmplicon-class}} object pre-processed
to have a sequenceTableMulti populated}

\item{mc.cores}{integer number of cores to use for parallelization}

\item{...}{paramter passed through to
\code{\link[dada2]{removeBimeraDenovo}}}
}
\value{
a code{\link{MulitAmplicon-class}} object with the
    sequenceTableNoChime filled
}
\description{
Remove chimeric sequencing read pairs from a MultiAmplcion
object.
}
\details{
This is a wrapper for the \code{\link[dada2]{removeBimeraDenovo}}
function of \code{dada2}. It works on an
\code{\link{MultiAmplicon-class}} object with the sequenceTable
slot filled. Use \code{\link{dadaMulti}},
\code{\link{derepMulti}}, \code{\link{mergeMulti}} and
\code{\link{sequenceTableMulti}} on a amplicon sorted (see
\code{\link{sortAmplicons}}) \code{\link{MultiAmplicon-class}}
object to preprocess your multi-marker data to this point.
}
\author{
Emanuel Heitlinger
}
