% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{PairedDada-class}
\alias{PairedDada-class}
\title{PairedDada-class}
\description{
A pair of two dada objects
}
\details{
dada-class objects as defined by the package
\code{\link[dada2]{dada2}} are bundeled as forward and reverse
read pairs in this object
}
\author{
Emanuel Heitlinger
}

