% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R, R/subset-methods.R
\docType{class}
\name{PairedDada-class}
\alias{PairedDada-class}
\alias{PairedDada}
\alias{length,PairedDada-method}
\alias{[,PairedDada,integer,missing,ANY-method}
\title{PairedDada-class}
\usage{
PairedDada(dadaF = list(), dadaR = list())

\S4method{length}{PairedDada}(x)

\S4method{[}{PairedDada,integer,missing,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{dadaF}{a dada object of forward reads}

\item{dadaR}{a dada object of reverse reads}

\item{x}{PairedDada-class object}

\item{i}{numeric to select}

\item{j}{not used}

\item{...}{not used}

\item{drop}{not used}
}
\description{
A pair of two dada objects
}
\details{
dada-class objects as defined by the package \code{dada2}
(function \code{\link[dada2]{dada}}) are bundeled as forward and
reverse read pairs in this object
}
\author{
Emanuel Heitlinger
}
