% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R, R/subset-methods.R
\docType{class}
\name{PairedDerep-class}
\alias{PairedDerep-class}
\alias{length,PairedDerep-method}
\alias{[,PairedDerep,index,missing,ANY-method}
\title{PairedDerep-class}
\usage{
\S4method{length}{PairedDerep}(x)

\S4method{[}{PairedDerep,index,missing,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{PairedDerep-class}

\item{i}{index, name or logical indicating which sequencing read file
pair to select or character string giving its name}

\item{j}{not used}

\item{...}{not used}

\item{drop}{not used}
}
\value{
A PairedDerep-class object
}
\description{
A pair of two derep objects
}
\details{
derep-class objects as defined by the package \code{dada2}
(\code{\link[dada2]{derepFastq}}] are bundled as forward and
reverse read pairs in this object
}
\section{Slots}{

\describe{
\item{\code{derepF}}{derep object containing forward read pairs created by
\code{dada2}'s \code{\link[dada2]{derepFastq}} function}

\item{\code{derepR}}{derep object containing reverse read pairs created by
\code{dada2}'s \code{\link[dada2]{derepFastq}} function}
}}

\author{
Emanuel Heitlinger
}
