% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R, R/subset-methods.R
\docType{class}
\name{PairedDerep-class}
\alias{PairedDerep-class}
\alias{length,PairedDerep-method}
\alias{[,PairedDerep,integer,missing,ANY-method}
\title{PiredDerep-class}
\usage{
\S4method{length}{PairedDerep}(x)

\S4method{[}{PairedDerep,integer,missing,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{PairedDerep-class}

\item{i}{numeric to select
##' @param j not used}

\item{...}{not used}

\item{drop}{not used}
}
\value{
A PairedDerep-class object
}
\description{
A pair of two derep objects
}
\details{
derep-class objects as defined by the package \code{dada2}
(\code{\link[dada2]{derepFastq}}] are bundeled as forward and
reverse read pairs in this object
}
\section{Slots}{

\describe{
\item{\code{derep}}{object containing forward read pairs created by
\code{dada2}'s \code{\link[dada2]{derepFastq}} function}

\item{\code{derepR}}{derep object containing reverse read pairscreated by
\code{dada2}'s \code{\link[dada2]{derepFastq}} function}
}}

\author{
Emanuel Heitlinger
}
