% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{PairedDerep-class}
\alias{PairedDerep-class}
\title{derep-class}
\value{
A PairedDerep-class object
}
\description{
A pair of two derep objects
}
\details{
derep-class objects as defined by the package
\code{\link[dada2]{derepFastq}} are bundeled as forward and
reverse read pairs in this object
}
\section{Slots}{

\describe{
\item{\code{derep}}{object containing forward read pairs created by
\code{dada2}'s \code{\link[dada2]{derepFastq}} function}

\item{\code{derepR}}{derep object containing reverse read pairscreated by
\code{dada2}'s \code{\link[dada2]{derepFastq}} function}
}}
\author{
Emanuel Heitlinger
}

