% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{PairedReadFileSet-class}
\alias{PairedReadFileSet-class}
\title{PairedReadFileSet-Class}
\usage{
## Constructors:
PairedReadFileSet(readsF, readsR)
}
\arguments{
\item{readsF}{The path and filenames of forward reads}

\item{readsR}{The path and filenames of reverse reads}
}
\value{
Constructors for defined classes
}
\description{
A class representing file paths to paired end reads.
}
\details{
Two character vectors of the same length specifying file names of
paired end reads can be stored in this class. Filenames a checked
for their existence.
}
\author{
Emanuel Heitlinger
}

