% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R, R/show-methods.R,
%   R/subset-methods.R
\docType{class}
\name{PairedReadFileSet-class}
\alias{PairedReadFileSet-class}
\alias{PairedReadFileSet}
\alias{length,PairedReadFileSet-method}
\alias{show,PairedReadFileSet-method}
\alias{[,PairedReadFileSet,ANY,ANY,ANY-method}
\title{A class representing file paths to paired end reads.}
\usage{
## Constructors:
PairedReadFileSet(readsF, readsR)

PairedReadFileSet(readsF, readsR, names = character())

\S4method{length}{PairedReadFileSet}(x)

\S4method{show}{PairedReadFileSet}(object)

\S4method{[}{PairedReadFileSet,ANY,ANY,ANY}(x, i)
}
\arguments{
\item{readsF}{The path and filenames containing forward (R1) reads}

\item{readsR}{The path and filenames containing reverse (R2) reads}
}
\value{
PairedReadFileSet
}
\description{
Two character vectors of the same length specifying file names of
paired end reads can be stored in this class. Filenames a checked
for their existence. Usually these are the filnames of quality
filtered fastq files already stratified into samples (one file
pair for each sample).
}
\section{Functions}{
\itemize{
\item \code{PairedReadFileSet}: Constructor for
PairedReadFileSet-class
}}

\section{Slots}{

\describe{
\item{\code{readF}}{A charcter vector specifying the file paths to files
containing forward (sometimes called R1) sequencing
reads. This vector can be named to store short short-name of
samples.}

\item{\code{readR}}{A character vector specifiying the file path to a file
containing reverse (sometimes called R2) sequencing
reads. This vector can be named to store short short-name of
samples. Names of readF and readR should be identical in this
case}
}}

\author{
Emanuel Heitlinger
}
