% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{PrimerPairsSet-class}
\alias{PrimerPairsSet}
\alias{PrimerPairsSet-class}
\title{A class representing sequences of forward and reverse
primers.}
\usage{
## Constructor:

PrimerPairsSet(primerF, primerR)

PrimerPairsSet(primerF, primerR)
}
\arguments{
\item{primerF}{Character vector or DNAStringSet. Can be named or
unnamed.}

\item{primerR}{Character vector or DNAStringSet of the same
length. Can be named or unnamed.}
}
\value{
PrimerPairsSet-class
}
\description{
The PrimerPairsSet class is a container for storing primer pairs.
This means exactly two \code{\link[Biostrings]{DNAStringSet}}
objects of the same length specifying primer-pairs. Primer
sequences can be provided as character strings and will be
converted to \code{\link[Biostrings]{DNAStringSet}} by the
constructor function of the same name. primerF and primerR have to
be of the same length to specify primer pairs
(primerF[i].primerR[i]; for i in 1:length(PrimerF)).  Warnings are
given if primer sequences are of unusal length (<16 or >26 bases).
}
\section{Functions}{
\itemize{
\item \code{PrimerPairsSet}: Constructor for
PrimerPairsSet-class
}}
\section{Slots}{

\describe{
\item{\code{primerF}}{DNAStringSet. Can be named or unnamed.}

\item{\code{primerR}}{DNAStringSet of the same length. Can be named or
unnnamed.}

\item{\code{names}}{Either constructed as a concatenation of names of
forward and reverse primers or of their sequences (if primer
sequences are unnamed)}

\item{\code{.mapF}}{(automatically generated) maps potentially duplicate
entries in FW primers to unique entries.}

\item{\code{.mapR}}{(auto-generated) maps potentially duplicate entries in
FW primers to unique entries.}

\item{\code{.uniqueF}}{(auto-generated) unique forward primers as
character strings.}

\item{\code{.uniqueR}}{(auto-generated) unique reverse primers as
character strings.}
}}
\author{
Emanuel Heitlinger
}
\seealso{
\code{\link[Biostrings]{DNAStringSet}}
}

