% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R, R/show-methods.R,
%   R/subset-methods.R
\docType{class}
\name{PrimerPairsSet-class}
\alias{PrimerPairsSet-class}
\alias{PrimerPairsSet}
\alias{length,PrimerPairsSet-method}
\alias{names,PrimerPairsSet-method}
\alias{show,PrimerPairsSet-method}
\alias{[,PrimerPairsSet,index,missing,ANY-method}
\title{A class representing sequences of forward and reverse primers.}
\usage{
PrimerPairsSet(primerF, primerR)

\S4method{length}{PrimerPairsSet}(x)

\S4method{names}{PrimerPairsSet}(x)

\S4method{show}{PrimerPairsSet}(object)

\S4method{[}{PrimerPairsSet,index,missing,ANY}(x, i, j, ..., drop = NA)
}
\arguments{
\item{primerF}{Character vector or DNAStringSet. Can be named or
unnamed.}

\item{primerR}{Character vector or DNAStringSet of the same
length. Can be named or unnamed.}

\item{x}{A \code{PrimerPairsSet-class} object.}

\item{i}{integer, or logical value which primer pairs to select or
character string giving the name of the primer pair}

\item{j}{not used}

\item{...}{not used}

\item{drop}{not used}

\item{x}{A \code{PrimerPairsSet-class} object.}
}
\value{
PrimerPairsSet-class
}
\description{
The PrimerPairsSet class is a container for storing primer pairs.
This means exactly two \code{\link[Biostrings]{DNAStringSet}}
objects of the same length specifying primer-pairs. Primer
sequences can be provided as character strings and will be
converted to \code{\link[Biostrings]{DNAStringSet}} by the
constructor function of the same name. primerF and primerR have to
be of the same length to specify primer pairs. Warnings are
given if primer sequences are of unusal length (<16 or >26 bases).

Accessor like functions
\code{length} gives the number of read paires in a
\code{PrimerPairsSet-class} object

\code{names} of primer-pairs (amplicons) in a
\code{PrimerPairsSet-class} object
}
\section{Slots}{

\describe{
\item{\code{primerF}}{DNAStringSet. Can be named or unnamed.}

\item{\code{primerR}}{DNAStringSet of the same length. Can be named or
unnnamed.}

\item{\code{names}}{Character string. Either constructed as a
concatenation of names of forward and reverse primers or of
their sequences (if primer sequences are unnamed).}

\item{\code{.mapF}}{(automatically generated) maps potentially duplicate
entries in FW primers to unique entries.}

\item{\code{.mapR}}{(auto-generated) maps potentially duplicate entries in
FW primers to unique entries.}

\item{\code{.uniqueF}}{(auto-generated) unique forward primers as
character strings.}

\item{\code{.uniqueR}}{(auto-generated) unique reverse primers as
character strings.}
}}

\seealso{
\code{\link[Biostrings]{DNAStringSet}}
}
\author{
Emanuel Heitlinger
}
