% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{methods}
\name{rawCounts}
\alias{rawCounts}
\alias{rawCounts,MultiAmplicon-method}
\title{rawCounts}
\usage{
rawCounts(x)

\S4method{rawCounts}{MultiAmplicon}(x)
}
\arguments{
\item{x}{A \code{\link{MultiAmplicon-class}} object}
}
\value{
A numerical matrix of counts for each amplicon
    (primer-pair) matched in each sample (paired read file)
}
\description{
Access rawCounts of a MultiAmplicon object
}
\details{
An accessor for the 'rawCounts' slot of a
\code{\link{MultiAmplicon-class}} object
}
\author{
Emanuel Heitlinger
}

