% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortAmplicons.R
\name{sortAmplicons}
\alias{sortAmplicons}
\alias{sortAmplicons,}
\alias{sortAmplicons-Method}
\title{sortAmplicons}
\usage{
sortAmplicons(MA, n = 1e+06, countOnly = FALSE, ...)
}
\arguments{
\item{MA}{MultiAmplicon-class object containing a set of paired
end files and a primer-pairs set.}

\item{n}{parameter passed to the yield functions of package
ShortRead. This controls the memory consumption during
streaming. Lower values result in lower memory requirements
but might result longer processing time due to more repeated
I/O operations reading the sequence files.}

\item{countOnly}{logical argument if set TRUE only a matrix of
read counts is returned}

\item{...}{addtional parameter so be passed to
Biostrings::isMatchingStartingAt. Be careful when using
multiple starting positions or allowing error. This could lead
to read pairs being assigned to multiple amplicons.}
}
\value{
MultiAmplicon: By default (countOnly=FALSE) a
    MultiAmplicon-class object is returned with the
    stratifiedFiles slot populated. Stratified file names are
    constructed using a unique string created by
    \code{\link{[base]tempfile}} and stored in R's
    \code{\link[base]{tempdir}}. If the countOnly is set only a
    numeric matrix of read counts is returned.
}
\description{
Sort different amplicons into a fully stratified samples x
amplicons structure based on primer matches.
}
\details{
This functions uses Biostrings::isMatchingStartingAt to match
primer sequences at the first postion of forward and reverse
sequences. These sequences are written to temporary files to allow
processing via standard metabarcoding pipelines (i.e. dada2 using
fastq streaming).
}
\author{
Emanuel Heitlinger
}

