% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortAmplicons.R
\name{sortAmplicons}
\alias{sortAmplicons}
\title{sortAmplicons}
\usage{
sortAmplicons(MA, ...)
}
\arguments{
\item{MA}{}

\item{...}{}
}
\value{
MultiAmplicon
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Emanuel Heitlinger
}

