% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toPhyloseq.R
\docType{methods}
\name{toPhyloseq,MultiAmplicon-method}
\alias{toPhyloseq,MultiAmplicon-method}
\title{toPhyloseq}
\usage{
\S4method{toPhyloseq}{MultiAmplicon}(MA, samples, ...)
}
\arguments{
\item{MA}{MultiAmplicon object with the \code{taxonTable} and
\code{sequenceTableNoChime} slots filled.}

\item{samples}{samples to include in phyloseq object}

\item{...}{additional arguments to be passed to
\code{\link[phyloseq]{phyloseq}}}
}
\value{
a \code{\link[phyloseq]{phyloseq}} object
}
\description{
Populate a phyloseq object with the contents of a MultiAmplicon
object
}
\details{
Information stored in a \code{MultiAmplicon} object is transferred
to a \code{\link[phyloseq]{phyloseq}} object.
}
\author{
Emanuel Heitlinger
}
