% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toPhyloseq.R
\name{toPhyloseq}
\alias{toPhyloseq}
\title{toPhyloseq}
\usage{
toPhyloseq(MA, samples, ...)
}
\arguments{
\item{MA}{MultiAmplicon object with the \code{taxonTable} and
\code{sequenceTableNoChime} slots filled.}

\item{samples}{samples to include in phyloseq object}

\item{...}{additional arguments to be passed to
\code{\link[phyloseq]{phyloseq}}}
}
\value{
a \code{\link[phyloseq]{phyloseq}} object
}
\description{
Create a phyloseq object from a MultiAmplicon object
}
\author{
Emanuel Heitlinger
}
